refact: ts
